// src/DockiUp.Frontend/src/app/api/model/gitCredentialDto.ts
/**
 * DockiUp.API | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GitCredentialDto {
    id: string;
    name: string;
    username: string;
    tokenMasked: string;
    projectCount: number;
    createdAt: string;
}